refactor(tests): helpers must die
