+++ test/utils/helpers.lua
@@ -127,62 +127,4 @@ function M.create_slnx_file(path, projects)
     return M.create_file(path, sln_string)
 end
 
----@return string?
-function M.predict_target(file_path, targets)
-    command("edit " .. vim.fs.joinpath(M.scratch, file_path))
-    return helpers.exec_lua(function(targets0)
-        local bufnr = vim.api.nvim_get_current_buf()
-        return require("roslyn.sln.utils").predict_target(bufnr, targets0)
-    end, targets)
-end
-
-function M.api_projects(target)
-    local sln = vim.fs.joinpath(M.scratch, target)
-    return helpers.exec_lua(function(target0)
-        return require("roslyn.sln.api").projects(target0)
-    end, sln)
-end
-
-function M.setup(config)
-    helpers.exec_lua(function(config0)
-        if config0.ignore_target then
-            local ignore = config0.ignore_target
-            config0.ignore_target = function(sln)
-                return string.match(sln, ignore) ~= nil
-            end
-        end
-
-        if config0.choose_target then
-            local choose = config0.choose_target
-            config0.choose_target = function(target)
-                return vim.iter(target):find(function(item)
-                    if string.match(item, choose) then
-                        return item
-                    end
-                end)
-            end
-        end
-
-        require("roslyn.config").setup(config0)
-    end, config)
-end
-
----Sets up a one-shot choose_target that picks a solution once, then clears itself.
----@param pattern string Pattern to match against solution paths
-function M.choose_solution_once(pattern)
-    helpers.exec_lua(function(pattern0)
-        local config = require("roslyn.config")
-        local current = config.get()
-
-        current.choose_target = function(targets)
-            -- Clear ourselves after being called once
-            current.choose_target = nil
-
-            return vim.iter(targets):find(function(item)
-                return string.match(item, pattern0)
-            end)
-        end
-    end, pattern)
-end
-
 return M
